requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docker==7.0.0
et-xmlfile==1.1.0
idna==3.6
influxdb-client==1.40.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
python-dateutil==2.8.2
pytz==2024.1
reactivex==4.0.4
requests==2.31.0
six==1.16.0
statsd-telegraf==3.2.1.post1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1

